% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{assembleMZQC}
\alias{assembleMZQC}
\title{Collects all 'mzQC' members from each entry in lst_qcMetrics and stores them in an overall mzQC object, which can be written to disk or augmented otherwise}
\usage{
assembleMZQC(lst_qcMetrics)
}
\arguments{
\item{lst_qcMetrics}{A list of qcMetric objects which have their mzQC member populated}
}
\description{
Collects all 'mzQC' members from each entry in lst_qcMetrics and stores them in an overall mzQC object, which can be written to disk or augmented otherwise
}
