% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bestKS}
\alias{bestKS}
\title{From a list of vectors, compute all vs. all Kolmogorov-Smirnoff p-values}
\usage{
bestKS(x)
}
\arguments{
  \item{x}{List of vectors, where each vector holds a
  distribution}
}
\value{
A data.frame with ks-test values of the "reference" to all
other distributions (see Details)
}
\description{
... and report the row of the matrix which has maximum sum
(i.e the best "reference" distribution). The returned
data.frame has as many rows as distributions given and two
columns. The first column 'name' gives the name of the list
element, the second column 'ks_best' gives the p-value of
the Kolmogorov-Smirnoff test to the "reference"
distribution (which was picked by maximising the sum of
p-values). Thus, the row with a p-value of 1 is the
reference distribution.
}

