% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_miscGGplot.R
\name{brewer.pal.Safe}
\alias{brewer.pal.Safe}
\title{Return color brew palettes, but fail hard if number of requested colors
is larger than the palette is holding.}
\usage{
brewer.pal.Safe(n = 3, palette = "Set1")
}
\arguments{
\item{n}{Number of colours}

\item{palette}{Name of palette (e.g. "set1")}
}
\value{
character vector of colors
}
\description{
Internally calls 'brewer.pal(n, palette)', checking 'n' beforehand.
}
