% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{checkEnglishLocale}
\alias{checkEnglishLocale}
\title{When MaxQuant is run with a wrong locale (i.e. the decimal separator is not a '.', but a ','), 
then MaxQuant results are plainly wrong and broken. The can be detected by, e.g. checking for negative charge annotation}
\usage{
checkEnglishLocale(df_evd)
}
\arguments{
\item{df_evd}{Evidence table from which we only need the 'charge' column}
}
\description{
When MaxQuant is run with a wrong locale (i.e. the decimal separator is not a '.', but a ','), 
then MaxQuant results are plainly wrong and broken. The can be detected by, e.g. checking for negative charge annotation
}
