% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_MQalign.R
\name{computeMatchRTFractions}
\alias{computeMatchRTFractions}
\title{Combine several data structs into a final picture for segmentation incurred by 'Match-between-runs'.}
\usage{
computeMatchRTFractions(qMBR, qMBRSeg_Dist_r, qMBRSeg_Dist_r_bg)
}
\arguments{
  \item{qMBR}{A data.frame as computed by
  peakSegmentation()}

  \item{qMBRSeg_Dist_r}{A data.frame as computed by
  inMatchWindow() for using the 'rtdiff' (=all) column}

  \item{qMBRSeg_Dist_r_bg}{A data.frame as computed by
  inMatchWindow() for using the 'rtdiff_bg' (=genuine only)
  column}
}
\value{
A data.frame which details the distribution of singlets and
pairs (inRT and outRT) for each Raw file and genuine vs.
all
}
\description{
...
}

