% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{createListOfPTXQCMetrics}
\alias{createListOfPTXQCMetrics}
\title{Create Html file with list of QC metrics}
\usage{
createListOfPTXQCMetrics(outdir = getwd(),
  outname = "PTXQC_list-of-metrics.html")
}
\arguments{
\item{outdir}{Target dir where PTXQC_list-of-metrics_template.html is written}

\item{outname}{Filename (without directory)}
}
\value{
Complete filename of written html file
}
\description{
Create Html file with list of QC metrics
}
