% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{CVDictionarySingleton}
\alias{CVDictionarySingleton}
\title{Define a Singleton class which can hold a CV dictionary (so we do not have to load the .obo files over and over again)}
\description{
Define a Singleton class which can hold a CV dictionary (so we do not have to load the .obo files over and over again)

Define a Singleton class which can hold a CV dictionary (so we do not have to load the .obo files over and over again)
}
\section{Super class}{
\code{\link[R6P:Singleton]{R6P::Singleton}} -> \code{CVDictionarySingleton}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Stores the data of the singleton. Set the data once before using the singleton all over the place}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{CVDictionarySingleton$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="R6P" data-topic="Singleton" data-id="initialize">}\href{../../R6P/html/Singleton.html#method-initialize}{\code{R6P::Singleton$initialize()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CVDictionarySingleton$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
