% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{darken}
\alias{darken}
\title{Make a color (given as name or in RGB) darker by factor x = [0 = black, 1=unchanged]}
\usage{
darken(color, factor = 0.8)
}
\arguments{
\item{color}{A color as understood by col2rgb}

\item{factor}{Between 0 (make black) and 1 (leave color as is)}
}
\value{
darkened color
}
\description{
Make a color (given as name or in RGB) darker by factor x = [0 = black, 1=unchanged]
}
