% Generated by roxygen2 (4.0.0): do not edit by hand
\name{delLCP}
\alias{delLCP}
\title{Removes the longest common prefix (LCP) from a vector of strings.}
\usage{
delLCP(x)
}
\arguments{
  \item{x}{Vector of strings with common prefix}
}
\value{
Shortened vector of strings
}
\description{
Removes the longest common prefix (LCP) from a vector of
strings.
}

