% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{delLCS}
\alias{delLCS}
\title{Removes the longest common suffix (LCS) from a vector of strings.}
\usage{
delLCS(x)
}
\arguments{
\item{x}{Vector of strings with common suffix}
}
\value{
Shortened vector of strings
}
\description{
Removes the longest common suffix (LCS) from a vector of strings.
}

