% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilenameMapper.R
\docType{class}
\name{FilenameMapper-class}
\alias{FilenameMapper-class}
\alias{FilenameMapper}
\title{Make sure to call
 
 $new()
 and then
 $readMappingFile(some_file)}
\description{
Make sure to call
 
 $new()
 and then
 $readMappingFile(some_file)
}
\section{Methods}{

\describe{
\item{\code{getShortNamesStatic(raw.files, max_len, fallbackStartNr = 1)}}{Static method: Shorten a set of Raw file names and return a data frame with the mappings.
 Mapping will have: $from, $to and optionally $best.effort (if shorting was unsuccessful and numbers had to be used)
 \itemize{
   \item{\verb{raw.files}  Vector of Raw files.}
   \item{\verb{max_len} Maximal length of shortening results, before resorting to canonical names (file 1,...).}
   \item{\verb{fallbackStartNr} Starting index for canonical names.}
 }
 \subsection{Return Value}{ data.frame with mapping.}}
}}

