% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findAlignReference}
\alias{findAlignReference}
\title{Return list of raw file names which were used as MaxQuant alignment reference.}
\usage{
findAlignReference(data, threshold = 1e-04)
}
\arguments{
  \item{data}{The data.frame with columns
  'retention.time.calibration' and 'raw.file'}

  \item{threshold}{Maximum range (min to max) of values
  required to be recognized as reference}
}
\value{
List of reference raw files (usually just one)
}
\description{
Usually there is only one reference -- except when
fractions are used. The reference can be identified by
using the 'retention.time.calibration' column in
evidence.txt. The raw.file with very small values is
usually the one (they are not exactly zero, even though
they should be!)
}
\details{
This function might return multiple, or no raw file names.
In this case the result should be treated with caution or
(better) regarded as failure.
}

