% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcMetric.R
\name{flattenList}
\alias{flattenList}
\title{Flatten lists of lists with irregular depths to just a list of items,
i.e. a list of the leaves (if you consider the input as a tree).}
\usage{
flattenList(x)
}
\arguments{
\item{x}{List of 'stuff' (could be lists or items or a mix)}
}
\value{
A flat list
}
\description{
Flatten lists of lists with irregular depths to just a list of items,
i.e. a list of the leaves (if you consider the input as a tree).
}
