% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gauss2Mix}
\alias{gauss2Mix}
\title{Given a vector of values sampled from a mixture of two Gaussians, compute the area contribution
of the higher/more abundant (=target) Gaussian.}
\usage{
gauss2Mix(x)
}
\arguments{
  \item{x}{Vector of numeric values}
}
\value{
Ratio of area of the target distribution vs. whole area
under the two Gaussians
}
\description{
Given a vector of values sampled from a mixture of two
Gaussians, compute the area contribution of the higher/more
abundant (=target) Gaussian.
}
\note{
This is pre-tuned for Match-time-differences
}

