% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{getCVDictionary}
\alias{getCVDictionary}
\title{Parse the content of 'qc-cv.obo' and 'psi-ms.obo' from the 'PTXQC/cv/' folder and return their union as ontology}
\usage{
getCVDictionary()
}
\value{
a list with 'id', 'name', 'def', 'parents', 'children' which contains the CV entries
}
\description{
Parse the content of 'qc-cv.obo' and 'psi-ms.obo' from the 'PTXQC/cv/' folder and return their union as ontology
}
