% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{getCVTemplate}
\alias{getCVTemplate}
\title{Fills a MzQCcvParameter object with id(accession) and name.
The value (if any) needs to be set afterwards.}
\usage{
getCVTemplate(accession, mzcv_dict)
}
\description{
Fills a MzQCcvParameter object with id(accession) and name.
The value (if any) needs to be set afterwards.
}
