% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{getECDF}
\alias{getECDF}
\title{Estimate the empirical density and return it}
\usage{
getECDF(samples, y_eval = (1:100)/100)
}
\arguments{
\item{samples}{Vector of input values (samples from the distribution)}

\item{y_eval}{Vector of points where CDF is evaluated (each percentile by default)}
}
\value{
Data.frame with columns 'x', 'y'
}
\description{
Estimate the empirical density and return it
}
\examples{
  plot(getECDF(rnorm(1e4)))
  
}
