% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQDataReader.R
\name{getFileEncoding}
\alias{getFileEncoding}
\title{Determine if a file is 'UTF-8' or 'UTF-8-BOM' (as of MQ2.4) or 'UTF-16BE' or 'UTF-16LE'}
\usage{
getFileEncoding(filename)
}
\arguments{
\item{filename}{Relative or absolute path to a file}
}
\value{
'' if the file does not exist or is not readable
}
\description{
Determine if a file is 'UTF-8' or 'UTF-8-BOM' (as of MQ2.4) or 'UTF-16BE' or 'UTF-16LE'
}
