% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_plots.R
\name{getHTMLTable}
\alias{getHTMLTable}
\title{Create an HTML table with an extra header row}
\usage{
getHTMLTable(data, caption = NA)
}
\arguments{
\item{data}{A data.frame which serves as table}

\item{caption}{A set of headlines, e.g. c("top line", "bottom line")}
}
\value{
table as html character string for cat()'ing into an html document
}
\description{
Create an HTML table with an extra header row
}
\examples{
  data = data.frame(raw.file = letters[1:4],
                    id.rate = 3:6)
  getHTMLTable(data, 
               caption = "some header line")

}
