% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_QCHeat.R
\name{getMetaData}
\alias{getMetaData}
\title{Extract meta information (orderNr, metric name, category)
from a list of Qc metric objects}
\usage{
getMetaData(lst_qcMetrics)
}
\arguments{
\item{lst_qcMetrics}{List of qcMetrics}
}
\value{
data.frame with columns 'name', 'order' and 'cat' (category)
}
\description{
Extract meta information (orderNr, metric name, category)
from a list of Qc metric objects
}

