% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{getMetricsObjects}
\alias{getMetricsObjects}
\title{Get all currently available metrics}
\usage{
getMetricsObjects(DEBUG_PTXQC = FALSE)
}
\arguments{
\item{DEBUG_PTXQC}{Use qc objects from the package (FALSE) or from environment (TRUE/DEBUG)}
}
\value{
List of matric objects
}
\description{
Get all currently available metrics
}
