% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{getProteinAndPeptideCounts}
\alias{getProteinAndPeptideCounts}
\title{Extract the number of protein groups and peptides observed per Raw file
from an evidence table.}
\usage{
getProteinAndPeptideCounts(d_evidence)
}
\arguments{
\item{d_evidence}{Data.frame of evidence.txt as read by MQDataReader}
}
\value{
Data.frame with columns 'proteinCounts', 'peptideCounts', 'category', 'MBRgain'
}
\description{
Required columns are "protein.group.ids", "fc.raw.file", "modified.sequence".
}
\details{
If match-between-runs was enabled during the MaxQuant run,
the data.frame returned will contain separate values for 'transferred' evidence
plus an 'MBRgain' column, which will give the extra MBR evidence in percent.
}

