% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{getProteinCounts}
\alias{getProteinCounts}
\title{Extract the number of protein groups observed per Raw file
from an evidence table.}
\usage{
getProteinCounts(df_evd)
}
\arguments{
\item{df_evd}{Data.frame of evidence.txt as read by MQDataReader}
}
\value{
Data.frame with columns 'fc.raw.file', 'counts', 'category', 'MBRgain'
}
\description{
Required columns are "protein.group.ids", "fc.raw.file" and "is.transferred".
}
\details{
If match-between-runs was enabled during the MaxQuant run,
the data.frame returned will contain separate values for 'transferred' evidence
plus an 'MBRgain' column, which will give the extra MBR evidence in percent.
}
