% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{getQualityMetricTemplate}
\alias{getQualityMetricTemplate}
\title{Fills a MzQCqualityMetric object with id(accession) and name.
The value (if any) and unit (if any) need to be set afterwards.}
\usage{
getQualityMetricTemplate(accession, mzcv_dict = CVDictionarySingleton$new())
}
\arguments{
\item{accession}{The ID (=accession) of the term in the CV}

\item{mzcv_dict}{A CV dictionary, as obtained by getCVDictionary(); defaults to a singleton, which needs to be filled manually beforehand}
}
\value{
An instance of MzQCqualityMetric
}
\description{
Fills a MzQCqualityMetric object with id(accession) and name.
The value (if any) and unit (if any) need to be set afterwards.
}
