% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{getReportFilenames}
\alias{getReportFilenames}
\title{Assembles a list of output file names, which will be created during reporting.}
\usage{
getReportFilenames(txt_folder)
}
\arguments{
\item{txt_folder}{Directory where the MaxQuant output resides}
}
\value{
List of output file names (just names :))
}
\description{
Assembles a list of output file names, which will be created during reporting.
}

