% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{getRunQualityTemplate}
\alias{getRunQualityTemplate}
\title{Get an mzQC runQuality without actual metrics, but with full metadata}
\usage{
getRunQualityTemplate(fc.raw.file, raw_file_mapping)
}
\arguments{
\item{fc.raw.file}{For which run}

\item{raw_file_mapping}{A data.frame with cols 'from', 'to' and maybe 'best.effort' (if shorting was unsuccessful), as e.g. obtained by a FilenameMapper$raw_file_mapping}
}
\value{
An MzQCrunQuality object
}
\description{
Get an mzQC runQuality without actual metrics, but with full metadata
}
