% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fcn_miscGGplot.R
\name{ggText}
\alias{ggText}
\title{Plot a text as graphic using ggplot2.}
\usage{
ggText(title, text, col = "black")
}
\arguments{
\item{title}{The title of the plot}

\item{text}{Centered text, can contain linebreaks}

\item{col}{Colour of text (excluding the title)}
}
\value{
ggplot object
}
\description{
Plot a text as graphic using ggplot2.
}

