% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_MQalign.R
\name{inMatchWindow}
\alias{inMatchWindow}
\title{Compute the fraction of 3D-peak pairs per Raw file which have an acceptable RT difference after alignment}
\usage{
inMatchWindow(data, colname, df.allowed.deltaRT)
}
\arguments{
  \item{data}{A data.frame with columns 'fc.raw.file' and
  !colname (param)}

  \item{colname}{Name of column which contains the RT
  difference of a pair}

  \item{df.allowed.deltaRT}{The allowed matching difference
  for each Raw file (as data.frame(fc.rawfile, m))}
}
\value{
A data.frame with one row for each raw.file and columns
'raw.file' and score 'withinRT' (0-1)
}
\description{
Using the result from 'idTransferCheck()', compute the
fraction of pairs which are within a certain RT tolerance.
}
\details{
Returned value is between 0 (bad) and 1 (all within
tolerance).
}

