% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lcpCount}
\alias{lcpCount}
\title{Count the number of chars of the longest common prefix}
\usage{
lcpCount(x)
}
\arguments{
  \item{x}{Vector of strings with common prefix}
}
\value{
Length of LCP
}
\description{
Count the number of chars of the longest common prefix
}

