% Generated by roxygen2 (4.0.0): do not edit by hand
\name{LCS}
\alias{LCS}
\title{Compute longest common substring of two strings.}
\usage{
LCS(s1, s2)
}
\arguments{
  \item{s1}{String one}

  \item{s2}{String two}
}
\value{
String containing the longest common substring
}
\description{
Implementation is very inefficient (dynamic programming in
R) --> use only on small instances
}

