% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{lcsCount}
\alias{lcsCount}
\title{Count the number of chars of the longest common suffix}
\usage{
lcsCount(x)
}
\arguments{
\item{x}{Vector of strings with common suffix}
}
\value{
Length of LCS
}
\description{
Count the number of chars of the longest common suffix
}
