% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{longestCommonPrefix}
\alias{longestCommonPrefix}
\title{Get the longest common prefix from a set of strings.}
\usage{
longestCommonPrefix(strings)
}
\arguments{
\item{strings}{Vector of strings}
}
\value{
Single string - might be empty ("")
}
\description{
Input is converted to character (e.g. from factor) first.
}
\examples{
  longestCommonPrefix(c("CBA.321", "CBA.77654", ""))    ## ""
  longestCommonPrefix(c("CBA.321", "CBA.77654", "CB"))  ## "CB"
  longestCommonPrefix(c("ABC.123", "ABC.456"))          ## "ABC."
  longestCommonPrefix(c("nothing", "in", "common"))     ## ""


}
