% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{longestCommonSuffix}
\alias{longestCommonSuffix}
\title{Like longestCommonPrefix(), but on the suffix.}
\usage{
longestCommonSuffix(strings)
}
\arguments{
\item{strings}{Vector of strings}
}
\value{
Single string - might be empty ("")
}
\description{
Like longestCommonPrefix(), but on the suffix.
}
\examples{

 longestCommonSuffix(c("123.ABC", "45677.ABC", "BC"))  ## "BC"
 longestCommonSuffix(c("123.ABC", "", "BC"))           ## ""
 longestCommonSuffix(c("123.ABC", "45677.ABC"))        ## ".ABC"
 longestCommonSuffix(c("nothing", "in", "common"))     ## ""
 
}
