% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modsToTable.R
\name{modsToTable}
\alias{modsToTable}
\title{Convert list of (mixed)modifications to a frequency table}
\usage{
modsToTable(mod_list)
}
\arguments{
\item{mod_list}{A vector with modifications, each for a specific peptide. Multiple mods per entry are allowed, each separated by comma.}
}
\value{
A data.frame with 'modification_names' and 'Freq' (0-100)
}
\description{
Convert list of (mixed)modifications to a frequency table
}
\examples{
modsToTable(c("Ox (M)", 
              "Unmodified",
              "Ox (M),Acetyl (Prot N-term)",
              "2 Ox (M)",
              "Unmodified",
              "Unmodified"))


}
