% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQDataReader.R
\name{MQDataReader$new}
\alias{MQDataReader$new}
\title{Constructor for class 'MQDataReader'.}
\description{
This class is used to read MQ data tables using readMQ() while holding
the internal raw file --> short raw file name mapping (stored in a member called 
'raw_file_mapping') and updating/using it every time readMQ() is called.
}

