% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQDataReader.R
\name{MQDataReader$readMappingFile}
\alias{MQDataReader$readMappingFile}
\title{Reads a mapping table of full Raw file names to shortened names.}
\arguments{
\item{filename}{Source filename to read.}
}
\value{
Returns \code{TRUE} if file was read, \code{FALSE} if it does not exist.
}
\description{
The internal structure \code{raw_file_mapping} is created using this file.
If the file is missing, nothing is done.
}
\details{
The file must have two columns named: 'orig.Name' and 'new.Name' and use Tab as separator.
I.e.
\preformatted{# This file can be used to manually substitute Raw file names within the report.
# The ordering of Raw files in the report can be changed by re-arranging the rows.
orig.Name  new.Name
2011_05_30_ALH_OT_21_VIL_TMT_FR01   myfile A
2011_05_30_ALH_OT_22_VIL_TMT_FR02   another B
}
}
