% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MQDataReader.R
\name{MQDataReader$substitute}
\alias{MQDataReader$substitute}
\title{Replaces values in the mq.data member with (binary) values.}
\arguments{
  \item{colname}{Name of the column (e.g. "contaminants")
  in the mq.data table}

  \item{valid_entries}{Vector of values to be replaced
  (must contain all values expected in the column -- fails
  otherwise)}

  \item{replacements}{Vector of values inserted with the
  same length as \code{valid_entries}.}
}
\value{
Returns \code{TRUE} if successful.
}
\description{
Most MQ tables contain columns like 'contaminants' or
'reverse', whose values are either empty strings or "+",
which is inconvenient and can be much better represented as
TRUE/FALSE. The params \code{valid_entries} and
\code{replacements} contain the matched pairs, which
determine what is replaced with what.
}

