% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCanalysisSoftware-class}
\alias{MzQCanalysisSoftware-class}
\alias{MzQCanalysisSoftware}
\title{Details of the software used to create the QC metrics}
\description{
Details of the software used to create the QC metrics
}
\section{Fields}{

\describe{
\item{\code{accession}}{Accession number identifying the term within its controlled vocabulary (pattern: ^[A-Z]+:[A-Z0-9]+$).}

\item{\code{name}}{Name of the controlled vocabulary term describing the software tool.}

\item{\code{version}}{Version number of the software tool.}

\item{\code{uri}}{Publicly accessible URI of the software tool or documentation.}

\item{\code{description}}{[optional] Definition of the controlled vocabulary term.}

\item{\code{value}}{[optional] Value of the software tool.}
}}


