% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCbaseQuality-class}
\alias{MzQCbaseQuality-class}
\alias{MzQCbaseQuality}
\title{Base class of runQuality/setQuality}
\description{
Base class of runQuality/setQuality
}
\section{Fields}{

\describe{
\item{\code{metadata}}{The metadata for this run/setQuality}

\item{\code{qualityMetrics}}{Array of MzQCqualityMetric objects}
}}


