% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCcontrolledVocabulary-class}
\alias{MzQCcontrolledVocabulary-class}
\alias{MzQCcontrolledVocabulary}
\title{An controlled vocabulary document, usually pointing to an .obo file}
\description{
An controlled vocabulary document, usually pointing to an .obo file
}
\section{Fields}{

\describe{
\item{\code{name}}{Full name of the controlled vocabulary.}

\item{\code{uri}}{Publicly accessible URI of the controlled vocabulary.}

\item{\code{version}}{[optional] Version of the controlled vocabulary.}
}}


\examples{
  MzQCcontrolledVocabulary$new(
    "Proteomics Standards Initiative Quality Control Ontology",
    "https://github.com/HUPO-PSI/qcML-development/blob/master/cv/v0_1_0/qc-cv.obo",
    "0.1.0")
  isValidMzQC(MzQCcontrolledVocabulary$new(
   "Proteomics Standards Initiative Quality Control Ontology"))
   
}
