% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCDateTime-class}
\alias{MzQCDateTime-class}
\alias{MzQCDateTime}
\title{An mzQC-formatted date+time, as required by the mzQC spec doc}
\description{
The format is "%Y-%m-%d %H:%M:%S".
}
\section{Fields}{

\describe{
\item{\code{datetime}}{A correctly formatted date time (use as read-only)}
}}


\examples{
   dt1 = MzQCDateTime$new("1900-01-01")
   dt2 = MzQCDateTime$new(Sys.time())
   ## test roundtrip conversion from/to JSON
   dt2$fromData(jsonlite::fromJSON(jsonlite::toJSON(dt1)))
   
}
