% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCinputFile-class}
\alias{MzQCinputFile-class}
\alias{MzQCinputFile}
\title{An inputfile within metadata for a run/setQuality}
\description{
An inputfile within metadata for a run/setQuality
}
\section{Fields}{

\describe{
\item{\code{name}}{The name MUST uniquely match to a location (specified below) listed in the mzQC file.}

\item{\code{location}}{Unique file location, REQUIRED to be specified as a URI. The file URI is RECOMMENDED to be publicly accessible.}

\item{\code{fileFormat}}{A MzQCcvParameter with 'accession' and 'name'.}

\item{\code{fileProperties}}{An array of MzQCcvParameter, usually with 'accession', 'name' and 'value'. Recommended are at least two entries: 
a) Completion time of the input file (MS:1000747) and b) Checksum of the input file (any child of: MS:1000561 ! data file checksum type).}
}}


