% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCmetadata-class}
\alias{MzQCmetadata-class}
\alias{MzQCmetadata}
\title{The metadata for a run/setQuality}
\description{
The metadata for a run/setQuality
}
\section{Fields}{

\describe{
\item{\code{label}}{Unique name for the run (for runQuality) or set (for setQuality).}

\item{\code{inputFiles}}{Array of MzQCinputFile objects}

\item{\code{analysisSoftware}}{Array of MzQCanalysisSoftware objects}

\item{\code{cvParameters}}{[optional] Array of cvParameters objects}
}}


