% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCmzQC-class}
\alias{MzQCmzQC-class}
\alias{MzQCmzQC}
\title{Root element of an mzQC document}
\description{
At least one of runQualities or setQualities MUST be present.
}
\section{Fields}{

\describe{
\item{\code{version}}{Version of the mzQC format.}

\item{\code{creationDate}}{Creation date of the mzQC file.}

\item{\code{contactName}}{Name of the operator/creator of this mzQC file.}

\item{\code{contactAddress}}{Contact address (mail/e-mail or phone)}

\item{\code{readMe}}{Description and comments about the mzQC file contents.}

\item{\code{runQualities}}{Array of MzQCbaseQuality;}

\item{\code{setQualities}}{Array of MzQCbaseQuality}

\item{\code{controlledVocabularies}}{Array of CV domains used (obo files)}
}}


