% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\docType{class}
\name{MzQCqualityMetric-class}
\alias{MzQCqualityMetric-class}
\alias{MzQCqualityMetric}
\title{The central class to store QC information}
\description{
The central class to store QC information
}
\section{Fields}{

\describe{
\item{\code{accession}}{Accession number identifying the term within its controlled vocabulary (pattern: ^[A-Z]+:[A-Z0-9]+$).}

\item{\code{name}}{Name of the controlled vocabulary element describing the metric.}

\item{\code{description}}{[optional] Definition of the controlled vocabulary term.}

\item{\code{value}}{[optional] Value of the metric (single value, n-tuple, table, matrix).
The structure is not checked by our mzQC implementation and must be handled by the caller}

\item{\code{unit}}{[optional] Array of unit(s), stored as MzQcvParameter}
}}


