% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabReader.R
\name{mzTab$get_df_evd}
\alias{mzTab$get_df_evd}
\title{Memberfunction to generate data.frame similar to MaxQuants evidence.txt from an mzTab object}
\description{
Usage: foo <- mzTab$new()
       foo <- foo$read.mzTab("file.mztab")
       metadata_from_file.mzTab <- foo$MTD
}
