% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzTabReader.R
\docType{class}
\name{mzTab-class}
\alias{mzTab-class}
\alias{mzTab}
\title{Class which holds the data from a .mzTab.}
\description{
Memberfunction read.mzTab() takes path to .mzTab file
Data from .mzTab sections is stored in seperate data.frames
}
\details{
Usage: foo <- mzTab$new()
       foo <- foo$read.mzTab("file.mztab")
       file.mztab <- foo$ALL
       metadata_from_file.mzTab <- foo$MTD
}

