% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabReader.R
\docType{class}
\name{MzTabReader-class}
\alias{MzTabReader-class}
\alias{MzTabReader}
\title{Class to read an mzTab file and store the tables internally.}
\description{
The 'sections' field is initialized after $readMzTab was called.
The 'fn_map' fields should be initialized via ...$fn_map$readMappingFile(...) manually if user-defined filename mappings are desired
and is automatically updated/queried when $readMzTab is called.
}
\section{Fields}{

\describe{
\item{\code{sections}}{MzTab sections as list. Valid list entries are: "MTD", "PRT", "PEP", "PSM", "SML", "filename" and "comments"}

\item{\code{fn_map}}{FilenameMapper which can translate raw filenames into something shorter}
}}

\section{Methods}{

\describe{
\item{\code{readMzTab(file)}}{Read a mzTab file into a list of 5 data.frames (one df per mzTab section).
Data.frames in the resulting list are named as follows:
  'MTD', 'PRT', 'PEP', 'PSM', 'SML',.
Additionally, 'filename' and 'comments' are valid list elements.}
}}

