% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\name{NULL_to_charNA}
\alias{NULL_to_charNA}
\title{converts a NULL to NA_character_; or returns the argument unchanged otherwise}
\usage{
NULL_to_charNA(char_or_NULL)
}
\arguments{
\item{char_or_NULL}{A string or NULL}
}
\description{
This is useful for missing list elements (which returns NULL), 
but when the missing element in refClass should be NA_character_ (and NULL would return an error)
}
\examples{
  NULL_to_charNA(NA)   ## NA
  NULL_to_charNA(NULL) ## NA_character_
  NULL_to_charNA("hi") ## "hi"
  
}
