% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzQC_classes.R
\name{NULL_to_NA}
\alias{NULL_to_NA}
\title{converts a NULL to NA; or returns the argument unchanged otherwise}
\usage{
NULL_to_NA(var_or_NULL)
}
\arguments{
\item{var_or_NULL}{A variable of any kind or NULL}
}
\description{
This is useful for missing list elements (which returns NULL), 
but when the missing element in refClass should be NA (and NULL would return an error)
}
\examples{
  NULL_to_NA(NA)   ## NA
  NULL_to_NA(NULL) ## NA
  NULL_to_NA("hi") ## "hi"
  
}
