% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ObjHandler$add}
\alias{ObjHandler$add}
\title{Add objects}
\arguments{
  \item{.}{A 'this' pointer. Use it to refer/change
  internal members. It's implicitly added, thus not
  required too call the function!}

  \item{obj}{Value to be inserted}

  \item{name}{String used as object name. Automatically
  generated if omitted.}

  \item{.}{A 'this' pointer. Use it to refer/change
  internal members. It's implicitly added, thus not
  required too call the function!}

  \item{name}{Name of list entry to return}
}
\value{
Always TRUE

The object
}
\description{
Also see 'getList'.

Also see 'getList'.
}

