% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ObjHandler$getList}
\alias{ObjHandler$getList}
\title{Get the objects as list}
\arguments{
  \item{.}{A 'this' pointer. Use it to refer/change
  internal members. It's implicitly added, thus not
  required too call the function!}
}
\value{
A list holding all objects
}
\description{
Also see 'addObj()'.
}

