% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{parseOBO}
\alias{parseOBO}
\title{Get the information of each CV term from an obo file.}
\usage{
parseOBO(cv_obo_file)
}
\arguments{
\item{cv_obo_file}{"xxx.obo"}
}
\value{
A list containing cv term information
}
\description{
Get the information of each CV term from an obo file.
}
