% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{pasten}
\alias{pasten}
\title{paste with newline as separator}
\usage{
pasten(...)
}
\arguments{
\item{...}{Arguments forwarded to paste()}
}
\value{
return value of paste()
}
\description{
paste with newline as separator
}
\examples{
  pasten("newline","separated")
  ## --> "newline\nseparated"
  
}
