% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_plots.R
\name{plot_ContsPG}
\alias{plot_ContsPG}
\title{Plot contaminants from proteinGroups.txt}
\usage{
plot_ContsPG(data)
}
\arguments{
\item{data}{A data.frame with columns 'group', 'cont_pc', 'logAbdClass'}
}
\value{
GGplot object
}
\description{
Plot contaminants from proteinGroups.txt
}
\examples{

 data = data.frame( 'group' = letters[1:10], 'cont_pc' = 2:11, 'logAbdClass' = c("low","high"))
 plot_ContsPG(data)

}
