% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_plots.R
\name{plot_IDRateBad}
\alias{plot_IDRateBad}
\title{Plot a table of Raw files which failed ID rate requirements.}
\usage{
plot_IDRateBad(data, title)
}
\arguments{
\item{data}{A data.frame with columns as described above}

\item{title}{Table title}
}
\value{
gTree object with class 'PTXQC_table'
}
\description{
Just to have their names explicitly in the report.
}
\details{
The input is a data.frame with two columns
  1st column - name of the Raw file
  2nd column - fraction of identified MS/MS spectra in percent
where each row represents one Raw file.
}
\examples{
  data = data.frame(raw.file = letters[1:4],
                    id.rate = 3:6)
  plot_IDRateBad(data, "Bad files")

}

