% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_plots.R
\name{plot_TopN}
\alias{plot_TopN}
\title{Plot line graph of TopN over Retention time.}
\usage{
plot_TopN(data)
}
\arguments{
\item{data}{A data.frame with columns 'fc.raw.file', 'scan.event.number', 'n'}
}
\value{
GGplot object
}
\description{
Number of Raw files must be 6 at most. Function will stop otherwise.
}
\examples{
  data = data.frame(fc.raw.file = rep(c("d","a","x"), each=10),
                    scan.event.number = 1:10,
                    n = 11:20)
  plot_TopN(data)

}
