% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_plots.R
\name{plotTableRaw}
\alias{plotTableRaw}
\title{Colored table plot.}
\usage{
plotTableRaw(data, colours = "black", fill = NA, just = "centre")
}
\arguments{
\item{data}{Table as Data.frame}

\item{colours}{Single or set of colours (col-wise)}

\item{fill}{Cell fill (row-wise)}

\item{just}{(ignored)}
}
\value{
gTable
}
\description{
Code taken from http://stackoverflow.com/questions/23819209/change-text-color-for-cells-using-tablegrob-in-r
}
