% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_plots.R
\name{print.PTXQC_table}
\alias{print.PTXQC_table}
\title{helper S3 class, enabling print(some-plot_Table-object)}
\usage{
\method{print}{PTXQC_table}(x)
}
\arguments{
\item{x}{Some Grid object to plot}
}
\value{
A function
}
\description{
helper S3 class, enabling print(some-plot_Table-object)
}

