% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_miscGGplot.R
\name{printWithFooter}
\alias{printWithFooter}
\title{Augment a ggplot with footer text}
\usage{
printWithFooter(gg_obj, bottom_left = NULL, bottom_right = NULL)
}
\arguments{
\item{gg_obj}{ggplot2 object to be printed}

\item{bottom_left}{Footer text for bottom left side}

\item{bottom_right}{Footer text for bottom right side}
}
\value{
-
}
\description{
Augment a ggplot with footer text
}
