% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{QCMetaFilenames}
\alias{QCMetaFilenames}
\title{Define a Singleton class which holds the full raw filenames (+path) and their PSI-MS CV terms for usage in the mzQC metadata}
\description{
The internal data is filled using, e.g. 'getMetaFilenames()'
}
\section{Super class}{
\code{\link[R6P:Singleton]{R6P::Singleton}} -> \code{QCMetaFilenames}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Stores the data of the singleton. Set the data once before using the singleton all over the place}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-QCMetaFilenames-clone}{\code{QCMetaFilenames$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="R6P" data-topic="Singleton" data-id="initialize"><a href='../../R6P/html/Singleton.html#method-Singleton-initialize'><code>R6P::Singleton$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-QCMetaFilenames-clone"></a>}}
\if{latex}{\out{\hypertarget{method-QCMetaFilenames-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{QCMetaFilenames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
