% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcMetric_MSMSScans.R
\docType{class}
\name{qcMetric_MSMSScans_TopNoverRT-class}
\alias{qcMetric_MSMSScans_TopNoverRT-class}
\alias{qcMetric_MSMSScans_TopNoverRT}
\title{Metric for msmsscans.txt, showing TopN over RT.}
\description{
Metric for msmsscans.txt, showing TopN over RT.
}

