% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qualBestKS}
\alias{qualBestKS}
\title{#
#Given a vector of values sampled from a mixture of two Gaussians, compute the area contribution
#of the higher/more abundant (=target) Gaussian.
#
#Returns a score between 1 (Target only) and 0 (Decoy only). Higher scores are better.
#We assume that the target distribution is much more peaked and narrow, whereas the decoy (false matches) are
#uniformly distributed across all bins (the Gaussian has a large SD).
#If there is not enough data, NA is returned.
#
#note This is pre-tuned for Match-time-differences
#
#param x Vector of numeric values
#param debug Set to TRUE to generate a plot, showing the fitted Gaussians to the data
#return Ratio of area of the target distribution vs. whole area under the two Gaussians
#
#importFrom mixtools normalmixEM2comp
#
#}
\usage{
qualBestKS(x)
}
\arguments{
  \item{x}{List of vectors, where each vector holds a
  distribution}
}
\value{
A data.frame with ks-test values of the "reference" to all
other distributions (see Details)
}
\description{
From a list of vectors, compute all vs. all
Kolmogorov-Smirnoff distance statistics (D)
}
\details{
... and report the row of the matrix which has maximum sum
(i.e the best "reference" distribution). The returned
data.frame has as many rows as distributions given and two
columns. The first column 'name' gives the name of the list
element, the second column 'ks_best' gives '1-statistic' of
the Kolmogorov-Smirnoff test to the "reference"
distribution (which was picked by maximising the sum of
'ks_best'). Thus, the row with a 'ks_best' of 1 is the
reference distribution.
}

