% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qualCenteredRef}
\alias{qualCenteredRef}
\title{Quality metric for 'centeredness' of a distribution with maximum reference}
\usage{
qualCenteredRef(x, tol)
}
\arguments{
  \item{x}{Position in interval [-tol, tol]}

  \item{tol}{Border of interval}
}
\value{
Value between [0, 1]
}
\description{
Ranges between 0 (worst score) and 1 (best score). The best
score is achieved when 'x' is close to the center of the
interval (given by 'tol'). If 'x' is close to the border
(on either side), the score decreases linearly to zero. Can
be used for uncalibrated mass errors, as a measure of how
well they are centered around 0.
}

