% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qualGauss}
\alias{qualGauss}
\title{Discete ChiSquare-test on raw (unbinned) data}
\usage{
qualGauss(x)
}
\arguments{
  \item{x}{Numeric vector of values (e.g. counts for charge
  state of 2)}
}
\value{
Numeric vector of qualities, same length of 'x', ranging
from 0 (bad agreement) to 1 (perfect)
}
\description{
Data are first assigned to bins of equal width using
'hist()'.
}
\details{
If only 'x' is given, its distribution of compared to a
uniform distribution. If 'y' is given as well, 'x' vs. 'y'
is compared (after binning). If counts of 'x' and 'y'
differ, the binned counts of the vector with higher counts
are rescaled by N/M to match the smaller one (this might
lead to bias).

Compute position of each element in a vector, assuming the
rest of the data is a Gaussian

For each index i, compute a Gaussian G from x \\ x_i and
assess the prob of x_i in G. The best result is 1, i.e. x_i
is in the center of all other values. The worst results is
0 (for very far outliers), i.e. the value is very far from
the median.
}

