% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qualGauss2Mix}
\alias{qualGauss2Mix}
\title{Given a vector of values sampled from a mixture of two Gaussians, compute the area contribution
of the higher/more abundant (=target) Gaussian.}
\usage{
qualGauss2Mix(x)
}
\arguments{
  \item{x}{Vector of numeric values}
}
\value{
Ratio of area of the target distribution vs. whole area
under the two Gaussians
}
\description{
Returns a score between 1 (Target only) and 0 (Decoy only).
Higher scores are better. We assume that the target
distribution is much more peaked and narrow, whereas the
decoy (false matches) are uniformly distributed across all
bins (the Gaussian has a large SD). If there is not enough
data, NA is returned.
}
\note{
This is pre-tuned for Match-time-differences
}

