% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_qualities.R
\name{qualGaussDev}
\alias{qualGaussDev}
\title{Compute probability of Gaussian (mu=m, sd=s) at a position 0, with reference
to the max obtainable probability of that Gaussian at its center.}
\usage{
qualGaussDev(mu, sd)
}
\arguments{
\item{mu}{Center of Gaussian}

\item{sd}{SD of Gaussian}
}
\value{
quality, ranging from 0 (bad agreement) to 1 (perfect, i.e. centered at 0)
}
\description{
Measure for centeredness around 0.
Highest score is 1, worst score is 0.
}
