% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_qualities.R
\name{qualLinThresh}
\alias{qualLinThresh}
\title{Quality metric with linear response to input, reaching the maximum score at the given threshold.}
\usage{
qualLinThresh(x, t = 1)
}
\arguments{
\item{x}{Numeric value(s) between [0, inf]}

\item{t}{Threshold value, which indicates 100\%}
}
\value{
Value between [0, 1]
}
\description{
Ranges between 0 (worst score) and 1 (best score).
Useful for performance measures where reaching a certain reference threshold 't'
will be enough to reach 100\%.
The input range from [0, t] is scored from 0-100\%.
}
