% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qualMBR}
\alias{qualMBR}
\title{Determine fraction of evidence which is wrongly inferred, i.e. was matched at different RT
even though it was already confirmed locally by MS/MS.}
\usage{
qualMBR(d_evd)
}
\arguments{
  \item{d_evd}{A data.frame of evidences containing the
  above columns}
}
\value{
A data.frame with one row per Raw file and two columns:
natively split peaks (%) and % of split peaks using
matching.
}
\description{
Sometimes, MQ split a feature into 2 or more if the
chromatograpic conditions are not optimal and there is a
drop in RT intensity. If both features contain successful
MS/MS scans, we will find the same peptide twice (with
slightly different RT) in the same charge state. This
constitutes a natively split peak and is rare (95% of all
peaks are unique).
}
\details{
If Match-between-runs is used and the RT alignment is not
perfect, then a peptide might be inferred at a wrong RT
position, even though this Raw file already contains MS/MS
evidence of this peptide. Usually the number of peak
duplicates rises drastically (e.g. only 75% of peaks are
unique after MBR was used). In most cases, the RT is too
far off to be a split peak. Its rather a lucky hit with
accidentally the same mass-to-charge, and thus the
intensity is random.

Required columns are 'match.time.difference', 'raw.file',
'modified.sequence', 'charge', 'type'.
}

