% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_qualities.R
\name{qualMedianDist}
\alias{qualMedianDist}
\title{Quality metric which measures the absolute distance from median.}
\usage{
qualMedianDist(x)
}
\arguments{
\item{x}{A vector numeric values between [0,1]}
}
\value{
A vector of the same size as x, with quality values between [0, 1]
}
\description{
Ranges between 0 (worst score) and 1 (best score).
Input must be between [0,1].
Deviations from the median of the sample represent the score for each sample point.
}
