% Generated by roxygen2 (4.0.0): do not edit by hand
\name{readMQ}
\alias{readMQ}
\title{Wrapper to read a MQ txt file (e.g. proteinGroups.txt).}
\usage{
readMQ(file, filter = "C+R", type = "pg", col_subset = NA,
  add_fs_col = T, LFQ_action = FALSE, ...)
}
\arguments{
  \item{file}{(relative) path to a MQ txt file ()}

  \item{filter}{Searched for "C" and "R". If present,
  [c]ontaminants and [r]everse hits are removed}

  \item{type}{Allowed values are: "pg" (proteinGroups)
  [default], adds abundance index columns (*AbInd*,
  replacing 'intensity') "sm" (summary), splits into three
  row subsets (raw.file, condition, total) Any other value
  will not add any special columns}

  \item{col_subset}{A vector of column names as read by
  read.delim(), e.g., spaces are replaced by dot already.
  If given, only columns with these names (ignoring
  lower/uppercase) will be returned (regex allowed) E.g.
  col_subset=c("^lfq.intensity.", "protein.name")}

  \item{add_fs_col}{If TRUE and a column 'raw.file' is
  present, an additional column 'fc.raw.file' will be added
  with common prefix AND common substrings removed
  (\code{\link{simplifyNames}}) E.g. two rawfiles named
  'OrbiXL_2014_Hek293_Control',
  'OrbiXL_2014_Hek293_Treated' will give 'Control',
  'Treated'}

  \item{LFQ_action}{[for type=='pg' only] An additional
  custom LFQ column ('cLFQ...') is created where zero
  values in LFQ columns are replaced by the following
  method IFF(!) the corresponding raw intensity is >0
  (indicating that LFQ is erroneusly 0) "toNA": replace by
  NA "impute": replace by lowest LFQ value >0 (simulating
  'noise')}

  \item{...}{Additional parameters passed on to
  read.delim()}
}
\value{
data frame
}
\description{
Since MaxQuant changes capitalization and sometimes even
column names, it seemed convenient to have a function which
just reads a txt file and returns unified column names,
irrespective of the MQ version. So, it unifies access to
columns (e.g. by using lower case for ALL columns) and
ensures columns are identically named across MQ versions.
We only of one case: "protease" == "enzyme".
}
\details{
If the file is empty, this function stops with an error.
}

