% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzTabReader.R
\name{readMzTab}
\alias{readMzTab}
\title{Read a mzTab file into a list of 5 data.frames (one df per mzTab section).}
\usage{
readMzTab(file)
}
\arguments{
\item{file}{A path to an mzTab file to read in.}
}
\description{
Data.frames in the resulting list are named as follows:
"MTD", "PRT", "PEP", "PSM", "SML"
}
\details{
Additionally, "filename" and "comments" are valid list elements.
}
