% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{removeSuffix}
\alias{removeSuffix}
\title{Removes the last suffix (including the last dot) from a filename.
If no dot exists, the full string is returned.}
\usage{
removeSuffix(filepath)
}
\arguments{
\item{filepath}{A filename (with optional path -- which is retained!)}
}
\value{
The input with removed suffix
}
\description{
Removes the last suffix (including the last dot) from a filename.
If no dot exists, the full string is returned.
}
\examples{
 removeSuffix("test.tar.gz")  # --> 'test.tar'
 removeSuffix("test.mzML")  # --> 'test'
 removeSuffix("/path/to/test.mzML")  # --> '/path/to/test'
 removeSuffix("test_no_dot")  # --> 'test_no_dot'

}
