% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{repEach}
\alias{repEach}
\title{Repeat each element x_i in X, n_i times.}
\usage{
repEach(x, n)
}
\arguments{
\item{x}{Values to be repeated}

\item{n}{Number of repeat for each x_i (same length as x)}
}
\value{
Vector with values from x, n times
}
\description{
Repeat each element x_i in X, n_i times.
}
\examples{
repEach(1:3, 1:3)  ## 1, 2, 2, 3, 3, 3
}

