% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ReportData.R
\name{ReportData$add}
\alias{ReportData$add}
\title{Add objects}
\arguments{
\item{.}{A 'this' pointer. Use it to refer/change internal members. It's implicitly added, thus not required too call the function!}

\item{obj}{Value to be inserted}

\item{type}{Type of object: metric, heatmap, params, name_mapping}

\item{name}{String used as object name. Automatically generated if omitted.}

\item{.}{A 'this' pointer. Use it to refer/change internal members. It's implicitly added, thus not required too call the function!}

\item{name}{Name of list entry to return}
}
\value{
Always TRUE (invisible)

The object
}
\description{
Also see 'getList'.

Also see 'getList'.
}

