% Generated by roxygen2 (4.0.1): do not edit by hand
\name{scale_x_discrete_reverse}
\alias{scale_x_discrete_reverse}
\title{Inverse the order of items on the x-axis (for discrete scales)}
\usage{
scale_x_discrete_reverse(values)
}
\arguments{
  \item{values}{The vector of values as given to the x
  aestetic.}
}
\value{
ggplot object, concatenatable with '+'
}
\description{
Inverse the order of items on the x-axis (for discrete
scales)
}

