% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_YAML.R
\name{setYAML}
\alias{setYAML}
\title{Set a YAML parameter to a certain value.}
\usage{
setYAML(config, param_name, value)
}
\arguments{
  \item{config}{A Yaml object as created by
  \code{\link[yaml]{yaml.load}}}

  \item{param_name}{A string which holds the param name,
  i.e. referring to a variable within 'config', e.g.
  "proteinGroup$doPlot"}

  \item{value}{New value to set}
}
\value{
The value
}
\description{
Set a YAML parameter to a certain value.
}
\note{
This function has a side-effect: it updates the parameter
passed as 'config' in the calling environment!!!
}

