% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{shortenStrings}
\alias{shortenStrings}
\title{Shorten a string to a maximum length and indicate shorting by appending '..'}
\usage{
shortenStrings(x, max_len = 12, print_warning = TRUE)
}
\arguments{
  \item{x}{Vector of input strings}

  \item{max_len}{Maximum length allowed}

  \item{print_warning}{Issue a warning which strings were
  shortened}
}
\value{
A vector of shortened strings
}
\description{
Some axis labels are sometimes just too long and printing
them will either squeeze the actual plot (ggplot) or make
the labels disappear beyond the margins (graphics::plot)
One ad-hoc way of avoiding this is to shorten the names,
hoping they are still meaningful to the viewer.
}
\examples{
r = shortenStrings(c("gamg_101", "gamg_101230100451", "jurkat_06_100731121305", "jurkat_06_1"))
all(r == c("gamg_101", "gamg_101230100..", "jurkat_06_1007..", "jurkat_06_1"))
}

