% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_miscGGplot.R
\name{theme_blank}
\alias{theme_blank}
\title{A blank theme (similar to the deprecated theme_blank())}
\usage{
theme_blank()
}
\description{
A blank theme (similar to the deprecated theme_blank())
}

