% Generated by roxygen2 (4.0.1): do not edit by hand
\name{thinOutBatch}
\alias{thinOutBatch}
\title{Apply 'thinOut' on all subsets of a data.frame, split by a batch column}
\usage{
thinOutBatch(data, filterColname, batchColname, binCount = 1000)
}
\arguments{
  \item{data}{The data.frame to be split and
  filtered(thinned)}

  \item{filterColname}{Name of the filter column as string}

  \item{batchColname}{Name of the split column as string}

  \item{binCount}{Number of bins in the 'filterColname'
  dimension.}
}
\description{
The binsize is computed from the global data range of the
filter column by dividing the range into binCount bins.
}

