% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_miscGGplot.R
\name{addGGtitle}
\alias{addGGtitle}
\title{Add title and subtitle to a ggplot}
\usage{
addGGtitle(pl, main, sub = "")
}
\arguments{
  \item{pl}{GGplot object}

  \item{main}{String for main title}

  \item{sub}{String for sub title}
}
\value{
A ggplot object
}
\description{
Found in http://www.antoni.fr/blog/?p=39 .. whewww...
modified a little
}

