% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{delLCS}
\alias{delLCS}
\title{Removes the longest common suffix (LCS) from a vector of strings.}
\usage{
delLCS(x)
}
\arguments{
\item{x}{Vector of strings with common suffix}
}
\value{
Shortened vector of strings
}
\description{
Removes the longest common suffix (LCS) from a vector of strings.
}
\examples{
delLCS(c("TK12345_H1"))                     ## ""
delLCS(c("TK12345_H1", "TK12345_H2"))       ## "TK12345_H1" "TK12345_H2" 
delLCS(c("TK12345_H1", "TK12!45_H1"))       ## "TK123"    "TK12!" 
 
}
