% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzQC.R
\name{getQualityMetricTemplate}
\alias{getQualityMetricTemplate}
\title{Fills a MzQCqualityMetric object with id(accession) and name.
The value (if any) and unit (if any) need to be set afterwards.}
\usage{
getQualityMetricTemplate(accession, mzcv_dict)
}
\description{
Fills a MzQCqualityMetric object with id(accession) and name.
The value (if any) and unit (if any) need to be set afterwards.
}
