% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%+\%}
\alias{\%+\%}
\title{A string concatenation function, more readable than 'paste()'.}
\usage{
a \%+\% b
}
\arguments{
  \item{a}{Char vector}

  \item{b}{Char vector}
}
\value{
Concatenated string (no separator)
}
\description{
A string concatenation function, more readable than
'paste()'.
}

