% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grepv}
\alias{grepv}
\title{Grep with values returned instead of indices}
\usage{
grepv(reg, data, ...)
}
\arguments{
  \item{reg}{regex param}

  \item{data}{container}

  \item{...}{other params forwarded to grep()}
}
\value{
values of data which matched the regex
}
\description{
Grep with values returned instead of indices
}
\examples{
grepv("x", c("abc", "xyz"))
  ## --> "xyz"
}

