% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{grepv}
\alias{grepv}
\title{Grep with values returned instead of indices.}
\usage{
grepv(reg, data, ...)
}
\arguments{
\item{reg}{regex param}

\item{data}{container}

\item{...}{other params forwarded to grep()}
}
\value{
values of data which matched the regex
}
\description{
The parameter 'value' should not be passed to this function since it is 
passed internally already.
}
\examples{
  grepv("x", c("abc", "xyz"))
  ## --> "xyz"
  
}
