% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_MQalign.R
\name{inMatchWindow}
\alias{inMatchWindow}
\title{For grouped peaks: separate them into in-width vs. out-width class.}
\usage{
inMatchWindow(data, df.allowed.deltaRT)
}
\arguments{
\item{data}{A data.frame with columns 'fc.raw.file', 'rtdiff_mixed', 'rtdiff_genuine'}

\item{df.allowed.deltaRT}{The allowed matching difference for each Raw file (as data.frame(fc.rawfile, m))}
}
\value{
A data.frame with one row for each raw.file and columns 'raw.file' and score 'withinRT' (0-1)
}
\description{
Looking at groups only: Compute the fraction of 3D-peak pair groups per Raw file which 
have an acceptable RT difference after alignment using the result from 'idTransferCheck()',
i.e. compute the fraction of groups which are within a certain RT tolerance.
}
\details{
Returned value is between 0 (bad) and 1 (all within tolerance).
}
