% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{lcpCount}
\alias{lcpCount}
\title{Count the number of chars of the longest common prefix}
\usage{
lcpCount(x)
}
\arguments{
  \item{x}{Vector of strings with common prefix}
}
\value{
Length of LCP
}
\description{
Count the number of chars of the longest common prefix
}

