% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lcsCount}
\alias{lcsCount}
\title{Count the number of chars of the longest common suffix}
\usage{
lcsCount(x)
}
\arguments{
  \item{x}{Vector of strings with common suffix}
}
\value{
Length of LCS
}
\description{
Count the number of chars of the longest common suffix
}

