% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MQDataReader$plotNameMapping}
\alias{MQDataReader$plotNameMapping}
\title{Plots the current mapping of Raw file names to their shortened version.}
\value{
Returns a mapping plot if mapping is available, 'NULL'
otherwise.
}
\description{
Convenience function to plot the mapping (e.g. to a PDF
device for reporting). The data frame can be accessed
directly via \code{.$raw_file_mapping}. If no mapping
exists, the function prints a warning to console and
returns NULL (which is safe to use in print(NULL)).
}

