% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MQDataReader.R
\name{MQDataReader$writeMappingFile}
\alias{MQDataReader$writeMappingFile}
\title{Writes a mapping table of full Raw file names to shortened names.}
\arguments{
  \item{filename}{Target filename to create.}
}
\value{
Returns \code{TRUE} if file was created, \code{FALSE} if it
already exists.
}
\description{
The internal structure \code{raw_file_mapping} is written
to the file specified. If the file already exists, nothing
is done.
}

