% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pasten}
\alias{pasten}
\title{paste with newline as separator}
\usage{
pasten(...)
}
\arguments{
  \item{...}{Arguments forwarded to paste()}
}
\value{
return value of paste()
}
\description{
paste with newline as separator
}
\examples{
pasten("newline","separated")
  ## --> "newline\\nseparated"
}

