% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pastet}
\alias{pastet}
\title{paste with tab as separator}
\usage{
pastet(...)
}
\arguments{
  \item{...}{Arguments forwarded to paste()}
}
\value{
return value of paste()
}
\description{
paste with tab as separator
}
\examples{
pastet("tab","separated")
  ## --> "tab\\tseparated"
}

