% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcMetric.R
\docType{class}
\name{qcMetric-class}
\alias{qcMetric}
\alias{qcMetric-class}
\title{Class which can compute plots (usually for a single metric).}
\description{
Reference class which is instanciated with a metric description and a
worker function (at initialization time, i.e. in the package)
and can produce plots (at runtime, when data is provided) using setData().
}
\section{Fields}{

\describe{
\item{\code{helpText}}{Description (lengthy) of the metric and plot elements}

\item{\code{workerFcn}}{Function which generates a result (usually plots). Data is provided using setData().}

\item{\code{plots}}{List of plots (after setData() was called)}

\item{\code{qcScores}}{[placeholder] Data.frame of scores from a qcMetric (computed within workerFcn())}

\item{\code{qcCat}}{[placeholder] QC category (LC, MS, or prep)}

\item{\code{qcName}}{[placeholder] Name of the qcScore in the heatmap}

\item{\code{heatmapOrder}}{[placeholder] column index during heatmap generation}
}}
\examples{

require(ggplot2)
dd = data.frame(x=1:10, y=11:20)
a = qcMetric$new(helpText="small help text", 
                 workerFcn=function(data, gtit) ## arbitrary arguments, matched during setData()
                 {
                   ## usually some code here to produce ggplots
                   pl = lapply(1:2, function(xx) ggplot(data) + geom_point(aes(x=x*xx,y=y)) + ggtitle(gtit))
                   return(list(plots = pl))
                 }, 
                 qcCat="LC", 
                 qcName="MS/MS Peak shape", 
                 heatmapOrder = 30)
## test some output
a$setData(dd, "my title")
a$plots  ## the raw plots
a$print(T) ## same as above
a$print(F) ## plots without title
a$getTitle()  ## get the title of the first plot (usually they are all identical)
a$helpText
a$qcName



}

