% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qualCentered}
\alias{qualCentered}
\title{Quality metric for 'centeredness' of a distribution}
\usage{
qualCentered(x)
}
\arguments{
  \item{x}{Numeric values (e.g. ppm errors)}
}
\value{
Value between [0, 1]
}
\description{
Ranges between 0 (worst score) and 1 (best score). Can be
used for calibrated mass errors, as a measure of how well
they are centered around 0.
}

