% Generated by roxygen2 (4.0.1): do not edit by hand
\name{qualGauss}
\alias{qualGauss}
\title{Compute position of each element in a vector, assuming the rest of the data is a Gaussian}
\usage{
qualGauss(x)
}
\arguments{
  \item{x}{Numeric vector of values (e.g. counts for charge
  state of 2)}
}
\value{
Numeric vector of qualities, same length of 'x', ranging
from 0 (bad agreement) to 1 (perfect)
}
\description{
For each index i, compute a Gaussian G from x \\ x_i and
assess the prob of x_i in G. The best result is 1, i.e. x_i
is in the center of all other values. The worst results is
0 (for very far outliers), i.e. the value is very far from
the median.
}

