% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQDataReader.R
\name{read.MQ}
\alias{read.MQ}
\title{Convenience wrapper for MQDataReader when only a single MQ file should be read
and file mapping need not be stored.}
\usage{
read.MQ(
  file,
  filter = "",
  type = "pg",
  col_subset = NA,
  add_fs_col = 10,
  LFQ_action = FALSE,
  ...
)
}
\arguments{
\item{file}{see \code{MQDataReader::readMQ()}}

\item{filter}{see \code{MQDataReader::readMQ()}}

\item{type}{see \code{MQDataReader::readMQ()}}

\item{col_subset}{see \code{MQDataReader::readMQ()}}

\item{add_fs_col}{see \code{MQDataReader::readMQ()}}

\item{LFQ_action}{see \code{MQDataReader::readMQ()}}

\item{...}{see \code{MQDataReader::readMQ()}}
}
\value{
see \code{MQDataReader::readMQ()}
}
\description{
For params, see \code{MQDataReader::readMQ()}.
}
