% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_misc.R
\name{thinOut}
\alias{thinOut}
\title{Thin out a data.frame by removing rows with similar numerical values in a certain column.}
\usage{
thinOut(data, filterColname, binsize)
}
\arguments{
\item{data}{The data.frame to be filtered}

\item{filterColname}{Name of the filter column as string}

\item{binsize}{Width of a bin}
}
\value{
Data.frame with reduced rows, but identical input columns
}
\description{
All values in the numerical column 'filterColname' are assigned to bins of width 'binsize'.
Only one value per bin is retained. All other rows are removed and the reduced
data frame will all its columns is returned.
}
