% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcn_YAML.R
\docType{class}
\name{YAMLClass-class}
\alias{YAMLClass}
\alias{YAMLClass-class}
\title{Query a YAML object for a certain parameter.}
\arguments{
\item{param_name}{A string which holds the param name, i.e. referring to a variable within 'config', e.g. "proteinGroup$doPlot"}

\item{default}{If seeked value is unknown, this will be used as new value}

\item{param_name}{A string which holds the param name, i.e. referring to a variable within 'config', e.g. "proteinGroup$doPlot"}

\item{value}{New value to set}

\item{filename}{File to write}

\item{yaml_obj}{A nested list object with configuration parameters for the report.
Useful to switch off certain plots or skip entire sections.
Will be converted via 'as.yaml()'}
}
\value{
The stored value (might be the default, if no value was present)

Set a YAML parameter to a certain value.

The value


Write YAML config (including some documentation) to a YAML file

TRUE on success
}
\description{
If the object has the param, then return it.
If the param is unknown, create it with the given default value and return the default.
}
\section{Fields}{

\describe{
\item{\code{yamlObj}}{A Yaml object as created by \code{\link[yaml]{yaml.load}}}
}}
\examples{
    yc = YAMLClass$new(list())
    val = yc$getYAML("cat$subCat", "someDefault")
    val  ## someDefault
    val = yc$setYAML("cat$subCat", "someValue")
    val  ## someValue
    yc$getYAML("cat$subCat", "someDefault") ## still 'someValue' (since its set already)

Query a YAML object for a certain parameter.

If the object has the param, then return it.
If the param is unknown, create it with the given default value and return the default.

}

